+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.0]
+
+Better error handling and logging.
+
+### Added
+
+- `catch2TestExplorer.defaultRunningTimeoutSec` to prevent infinite loops
+
+
 ## [2.2.6] - 2018-12-10
 
 Stability improvements.